Umstellung Double zu Float LOGGER_CONT: Werte in Reihen zu 8 bytes sichern MessungDataStringArray neu: Array der Messungen

+++ Power_Interface/diagramm.swift
@@ -121,7 +121,7 @@ static let minorrasterhorizontal = 10
    }
 
    
-   open func setWerteArray(werteArray:[Double])
+   open func setWerteArray(werteArray:[Float])
    {
  //     Swift.print("")
       let AnzeigeFaktor:CGFloat = 1.0//= maxSortenwert/maxAnzeigewert;
